+++ tests/xml/test_constant.py
@@ -5,7 +5,6 @@ import os
 import sys
 import unittest
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../../src')))
-#import autosar.xml.enumeration as ar_enum # noqa E402
 import autosar.xml.element as ar_element # noqa E402
 import autosar # noqa E402